css refactoring

+++ spring_server/src/main/java/com/example/ekstraklasa/controllers/DashboardUserController.java
@@ -3,7 +3,6 @@ package com.example.ekstraklasa.controllers;
 import com.example.ekstraklasa.models.Team;
 import com.example.ekstraklasa.models.Users;
 import com.example.ekstraklasa.services.FavouriteTeamService;
-import com.example.ekstraklasa.services.MatchService;
 import com.example.ekstraklasa.services.TeamService;
 import com.example.ekstraklasa.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,15 +22,12 @@ public class DashboardUserController {
     private final UserService userService;
     private final TeamService teamService;
     private final FavouriteTeamService favouriteTeamService;
-    private final MatchService matchService;
 
     @Autowired
-    public DashboardUserController(UserService userService, TeamService teamService,
-                                   FavouriteTeamService favouriteTeamService, MatchService matchService) {
+    public DashboardUserController(UserService userService, TeamService teamService, FavouriteTeamService favouriteTeamService) {
         this.userService = userService;
         this.teamService = teamService;
         this.favouriteTeamService = favouriteTeamService;
-        this.matchService = matchService;
     }
 
     @RequestMapping(value = "/favourite_teams", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -42,11 +38,12 @@ public class DashboardUserController {
         return new ResponseEntity<>(response, status);
     }
 
-    @RequestMapping(value = "/favourite_teams", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
+    @RequestMapping(value = "/favourite_teams", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> saveUsersFavouriteTeams(@RequestBody String object) {
         Users user = userService.findUserByUsername(Users.getUserName());
         Optional<List<Team>> teams = teamService.getAllByName(object);
-        favouriteTeamService.deleteAllByUserId(user.getId());
+        if(teams.isPresent())
+            favouriteTeamService.deleteAllByUserId(user.getId());
         Map<String,Object> response = favouriteTeamService.saveAllByUserId(teams, user);
         HttpStatus status = response.get("Status").equals(200) ? HttpStatus.OK : HttpStatus.INTERNAL_SERVER_ERROR;
         return new ResponseEntity<>(response, status);
